refactor: post management
